# Cache the next position of `urlreplacer.prefix`