Improved performance due to `bytes.Index` repeated calls.

Fixes #5942