@bep 2019-08-06
Thanks. The relevant benchmarks before/after this PR:

```bash
benchmark                      old ns/op     new ns/op     delta
BenchmarkAbsURL-4              8347          7225          -13.44%
BenchmarkAbsURLSrcset-4        7768          5518          -28.96%
BenchmarkXMLAbsURLSrcset-4     7656          5378          -29.75%
BenchmarkXMLAbsURL-4           3554          2997          -15.67%

benchmark                      old allocs     new allocs     delta
BenchmarkAbsURL-4              16             16             +0.00%
BenchmarkAbsURLSrcset-4        23             23             +0.00%
BenchmarkXMLAbsURLSrcset-4     21             21             +0.00%
BenchmarkXMLAbsURL-4           8              8              +0.00%

benchmark                      old bytes     new bytes     delta
BenchmarkAbsURL-4              1570          1570          +0.00%
BenchmarkAbsURLSrcset-4        1289          1289          +0.00%
BenchmarkXMLAbsURLSrcset-4     1368          1368          +0.00%
BenchmarkXMLAbsURL-4           929           930           +0.11%
```